+++ compiler-core/src/docs/tests.rs
@@ -1,7 +1,10 @@
 use std::time::SystemTime;
 
 use crate::{
-    build::{Mode, NullTelemetry, PackageCompiler, StaleTracker, TargetCodegenConfiguration},
+    build::{
+        Mode, ModulesCompilation, NullTelemetry, PackageCompiler, StaleTracker,
+        TargetCodegenConfiguration,
+    },
     config::PackageConfig,
     io::{memory::InMemoryFileSystem, FileSystemWriter},
     paths::ProjectPaths,
@@ -93,8 +96,17 @@ fn compile(config: PackageConfig, modules: Vec<(&str, &str)>) -> EcoString {
     let build = root.join("build");
     let lib = root.join("lib");
     let paths = ProjectPaths::new(root.clone());
-    let mut compiler =
-        PackageCompiler::new(&config, Mode::Dev, &root, &build, &lib, &target, ids, fs);
+    let mut compiler = PackageCompiler::new(
+        &config,
+        Mode::Dev,
+        ModulesCompilation::CompileAll,
+        &root,
+        &build,
+        &lib,
+        &target,
+        ids,
+        fs,
+    );
     compiler.write_entrypoint = false;
     compiler.write_metadata = false;
     compiler.compile_beam_bytecode = true;