+++ compiler-wasm/src/lib.rs
@@ -6,7 +6,8 @@ mod wasm_filesystem;
 use camino::Utf8PathBuf;
 use gleam_core::{
     build::{
-        Mode, NullTelemetry, PackageCompiler, StaleTracker, Target, TargetCodegenConfiguration,
+        Mode, ModulesCompilation, NullTelemetry, PackageCompiler, StaleTracker, Target,
+        TargetCodegenConfiguration,
     },
     config::PackageConfig,
     io::{FileSystemReader, FileSystemWriter},
@@ -193,6 +194,7 @@ fn do_compile_package(project: Project, target: Target) -> Result<(), Error> {
     let mut compiler = PackageCompiler::new(
         &config,
         Mode::Dev,
+        ModulesCompilation::CompileAll,
         &package,
         &out,
         &lib,